server/requirements.txt
anyio==3.5.0
asgiref==3.4.1
autopep8==1.6.0
click==8.0.3
fastapi==0.72.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
psycopg2==2.9.3
pycodestyle==2.8.0
pycryptodome==3.12.0
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
starlette==0.17.1
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
